add `/scan/<scan_id>/actions/add-workers` - 2
